+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.9.0 (2020-07-23)
+
 ### Added
 
 -   `conditionallyExecutePredicates` to `FormSubmissionEvent`